naive fix for goTo bug with multiple video completions
We can check the video source on end to be sure it is the same, however if the video happens to have the same source from a second invocation of that source (maybe two different nodes play the same video) then every registered async will fire when that video ends. The first set will not have been properly cancelled. I have called attention to this already in issue # 9

I have added a skipped test for this bug